requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
iniconfig==2.0.0
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
pydantic==2.7.0
pydantic-core==2.18.1
PyMySQL==1.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tomli==2.0.1
typing-extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
zipp==3.18.1
